# claim/api-docs.yaml
openapi: 3.1.0
info:
  title: CargoTech Claim API
  description: "Реестр претензий, расчёты, версии текста, комментарии и статусы"
  version: v1
servers:
- url: http://localhost:8083
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/claims:
    get:
      tags:
      - claim-controller
      operationId: list_3
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - PENDING_LEGAL_REVIEW
          - LEGAL_APPROVED
          - SENT
          - AWAITING_RESPONSE
          - PAID
          - ESCALATED_TO_COURT
          - CANCELLED
          - CLOSED_IN_COURT
      - name: creditorId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: debtorId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: assignedLawyerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseClaimListItemResponse"
    post:
      tags:
      - claim-controller
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
...
  /api/v1/claims/{claimId}/submit-to-legal-review:
    post:
      tags:
      - claim-controller
      operationId: submitToLegalReview
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/confirm-non-payment:
    post:
      tags:
      - claim-controller
      summary: Подтвердить отсутствие оплаты без изменения статуса претензии
      operationId: confirmNonPayment
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/send:
    post:
      tags:
      - claim-controller
      operationId: send
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/mark-paid:
    post:
      tags:
      - claim-controller
      operationId: markPaid
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/escalate-to-court:
    post:
      tags:
      - claim-controller
      operationId: escalateToCourt
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
...
  /api/v1/claims/{claimId}/close-in-court:
    post:
      tags:
      - claim-controller
      operationId: closeInCourt
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/cancel:
    post:
      tags:
      - claim-controller
      operationId: cancel
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
  /api/v1/claims/{claimId}/approve:
    post:
      tags:
      - claim-controller
      operationId: approve
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
...
  /api/v1/claims/{claimId}:
    get:
      tags:
      - claim-controller
      operationId: get_3
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
    delete:
      tags:
      - claim-controller
      operationId: delete_2
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
    patch:
      tags:
      - claim-controller
      operationId: update_3
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClaimRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimDetailsResponse"
...
  /api/v1/claims/{claimId}/status-history:
    get:
      tags:
      - claim-controller
      operationId: statusHistory
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusHistoryResponse"
  /api/v1/calculations/claim/{claimId}:
    get:
      tags:
      - calculation-controller
      operationId: getLatest
      parameters:
      - name: claimId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClaimCalculationResponse"
components:
  schemas:
...
    CreateClaimRequest:
      type: object
      properties:
        claimNumber:
          type: string
          maxLength: 128
          minLength: 0
        shipmentId:
          type: string
          format: uuid
        creditorId:
          type: string
          format: uuid
        debtorId:
          type: string
          format: uuid
        claimType:
          type: string
          enum:
          - PAYMENT_DELAY
        reason:
          type: string
        principalDebt:
          type: number
        penaltyAmount:
          type: number
        assignedLawyerId:
          type: string
          format: uuid
        nonPaymentConfirmed:
          type: boolean
        nonPaymentConfirmationComment:
          type: string
        draftContent:
          type: string
      required:
      - shipmentId
    ClaimDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        claimNumber:
          type: string
        claimType:
          type: string
          enum:
          - PAYMENT_DELAY
        status:
          type: string
          enum:
          - DRAFT
          - PENDING_LEGAL_REVIEW
          - LEGAL_APPROVED
          - SENT
          - AWAITING_RESPONSE
          - PAID
          - ESCALATED_TO_COURT
          - CANCELLED
          - CLOSED_IN_COURT
        reason:
          type: string
        shipmentId:
          type: string
          format: uuid
        shipmentNumber:
          type: string
        contractId:
          type: string
          format: uuid
        contractNumber:
          type: string
        creditorId:
          type: string
          format: uuid
        creditorName:
          type: string
        debtorId:
          type: string
          format: uuid
        debtorName:
          type: string
        principalDebt:
          type: number
        penaltyAmount:
          type: number
        totalAmount:
          type: number
        nonPaymentConfirmed:
          type: boolean
        nonPaymentConfirmedAt:
          type: string
          format: date-time
        nonPaymentConfirmedBy:
          type: string
          format: uuid
        nonPaymentConfirmationComment:
          type: string
        assignedLawyerId:
          type: string
          format: uuid
        finalVersionId:
          type: string
          format: uuid
        lastPaymentCheckId:
          type: string
          format: uuid
        approvedAt:
          type: string
          format: date-time
        approvedBy:
          type: string
          format: uuid
        sentAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        cancellationReasonCode:
          type: string
        cancellationReason:
          type: string
        escalatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    StatusChangeRequest:
      type: object
      properties:
        reason:
          type: string
        cancellationReasonCode:
          type: string
...
    ClaimCalculationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimId:
          type: string
          format: uuid
        calculationVersion:
          type: integer
          format: int32
        principalDebt:
          type: number
        paidAmount:
          type: number
        remainingDebt:
          type: number
        overdueStartDate:
          type: string
          format: date
        calculationDate:
          type: string
          format: date
        overdueDays:
          type: integer
          format: int32
        penaltyType:
          type: string
          enum:
          - CONTRACT_PENALTY
          - ARTICLE_395
          - NONE
        penaltyRate:
          type: number
        penaltyAmount:
          type: number
        totalAmount:
          type: number
        formula:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    UpdateClaimRequest:
      type: object
      properties:
        reason:
          type: string
        principalDebt:
          type: number
        penaltyAmount:
          type: number
        assignedLawyerId:
          type: string
          format: uuid
        nonPaymentConfirmed:
          type: boolean
        nonPaymentConfirmationComment:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    ClaimListItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimNumber:
          type: string
        claimType:
          type: string
          enum:
          - PAYMENT_DELAY
        status:
          type: string
          enum:
          - DRAFT
          - PENDING_LEGAL_REVIEW
          - LEGAL_APPROVED
          - SENT
          - AWAITING_RESPONSE
          - PAID
          - ESCALATED_TO_COURT
          - CANCELLED
          - CLOSED_IN_COURT
        shipmentId:
          type: string
          format: uuid
        shipmentNumber:
          type: string
        creditorId:
          type: string
          format: uuid
        creditorName:
          type: string
        debtorId:
          type: string
          format: uuid
        debtorName:
          type: string
        assignedLawyerId:
          type: string
          format: uuid
        principalDebt:
          type: number
        penaltyAmount:
          type: number
        totalAmount:
          type: number
        overdueDays:
          type: integer
          format: int32
        paymentDays:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PageResponseClaimListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ClaimListItemResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
    StatusHistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimId:
          type: string
          format: uuid
        previousStatus:
          type: string
          enum:
          - DRAFT
          - PENDING_LEGAL_REVIEW
          - LEGAL_APPROVED
          - SENT
          - AWAITING_RESPONSE
          - PAID
          - ESCALATED_TO_COURT
          - CANCELLED
          - CLOSED_IN_COURT
        newStatus:
          type: string
          enum:
          - DRAFT
          - PENDING_LEGAL_REVIEW
          - LEGAL_APPROVED
          - SENT
          - AWAITING_RESPONSE
          - PAID
          - ESCALATED_TO_COURT
          - CANCELLED
          - CLOSED_IN_COURT
        reason:
          type: string
        changedBy:
          type: string
          format: uuid
        changedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT